tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/send/011.html

<!doctype html>
<title>WebSockets: sending non-ascii, combining chars and non-BMP</title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>

async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  ws.onopen = t.step_func(function(e) {
    ws.send('\u00E5 a\u030A \uD801\uDC7E');
  });
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, '\u00E5 a\u030A \uD801\uDC7E');
    ws.onclose = t.step_func(function(e) {
      ws.onclose = t.unreached_func();
      t.step_timeout(() => t.done(), 50);
    })
    ws.close();
  })
  ws.onclose = ws.onerror = t.unreached_func();
});
</script>
